Update header navigation and recent donations fetching
Added 'Donors' link to the header navigation and reflected it in active page styling. The 'Applications' link was also moved to maintain the relevance in navigation order. For the external 'donate' link, opted for usage of variable 'external_donation_url' for better code manageability and future update facilitation.
In 'WelcomeController', the fetching of recent donations is refactored to use a method 'recents' on the Donation model to make the controller look cleaner and to encapsulate the functionality in the model, where it is more appropriate.