Remove "Donations" from admin panel
All files related to the Donations in the admin panel have been removed as per the requirements. On the other hand, a new functionality of adding multiple donations while creating or editing a donor has been implemented. This has been done to centrally manage the donations from the Donor module itself. For this, a new partial "donation_fields" has been created and used in the Donor form. The Donation model now belongs to a Donor from the User model. Furthermore, small changes like changing the wrong label of "Web Address" and incorrect role default value of "guest" to "subscriber" were made.