Remove application related files and controllers
This commit removes all application related files, models, and controllers.
This is part of a larger refactoring effort, wherein we are moving away from the previous application system to a new, more modular design. As such, all existing application-related code has been made redundant and is being removed to improve maintainability and to prepare the codebase for the new implementation.