Remove question-related models, views, and migrations
Removed the question-related files (controllers, models, views, tests, and migrations) as the structure for handling questions was redesigned. With this restructuring, questions are now managed as part of the 'Grant' model rather than in a standalone 'Question' model. This was done to simplify the application, minimize potential points of failure, and enhance conceptual de-coupling.