Create Grants and Grant Requests structure and refactor submit button classes
The commit adds new views, models and migrations to create 'Grants' and 'Grant Requests' functionality in the admin area. It also updates several other views changing the css class of form submit buttons from a specific style to a general 'btn-primary' class. This was done to ensure button styling is consistent throughout the application and makes future style changes easier to manage.