Swap link destinations and rebrand Donors to Supporters
Updated links in the navigation menus to swap the destinations of "Board" and "Donors". "Donors" was renamed to "Supporters" in several locations across the app to more accurately reflect the function and purpose of these individuals and organizations. Additionally, the inline 'Donate' link was commented out in several places, possibly in anticipation of future reinstatement. The session timeout in sorcery.rb was also set to a specific value for security reasons. Lastly, details on the footer and title headlining were updated for accuracy and coherence.