Update navigation, donor display, mailing settings and forms
Updated menu items in header and footer for improved user experience by rearranging and reenabling certain selections. The donor form was enhanced to include a 'level'field and present the total donation of a supporter, aiding in better organisation of data. The setting for mailing to board members and applicants on grant submission got set up under development configuration. The 'phone' field was also appended in the grant requests form to allow applicants to provide contact details. The way donors display on index was reshaped, presenting them based on their giving levels. Displaying donor level and donation history on the 'show' donor page allows in-depth examination of individual contributors.