Add total donation amount to donor admin panel
The total donated amount is now displayed on the admin panel for donors. This was done by modifying the donor's controller to calculate the total amount donated. This calculation was then inserted into the 'index.html.erb' file to be displayed on the admin panel. This addition provides a quick overview to admins of the total amount donated.