Add GrantDecisions feature and handle exceptions in GrantRequests
This commit introduces the GrantDecisions model and controller, allowing the decision about a grant request to be created, updated and persisted in the database. This includes the decision status and the amount awarded.

Additionally, this modifies a number of existing models, views, and controllers including GrantRequests, Donors, and mailer services, notably adding exception handling for failed mailer service in the GrantRequests controller. Among the alterations, 'alumni' and 'how_heard' were added as store_accessors in the Donor model.

This also updates the stylesheets and HTML of some forms and show pages.

These changes provide functionality to make a decision about a grant request and give more detailed information about donors. Furthermore, the exception handling offers a more robust way of dealing with system failures.