Improve grant request decision visibility & form enhancements
The decision status for grant requests in the admin view has been revised for better clarity. Prior to this change, the decision status for approved or denied grants was not explicitly displayed. We've added conditional logic to check if a decision has been made, and if so, whether it was approved or denied. This status is then displayed accordingly.

Additionally, form inputs in 'grant reviews' and 'grant decisions' views have been modified for a more user-friendly interface, with clearer labels and improved text contrast for dark mode.