Update gems and improve grant requests functionality
This commit includes an update to several gems and includes major improvements to the grant requests functionality. These updates include the incorporation of UserSearchable in GrantRequestsController for a filter functionality, various changes in the mailer and views files for grant request email, and adding new methods in the GrantRequest model to improve filtering. Improvements are also made for views in the GrantRequest index page and partial html files, improving application readability and functionalities.