Add Awardees page and update Ruby version
This commit includes creation of a new Awardees page displaying grant recipients. The Ruby version has also been updated. New features include a controller for awardees, along with corresponding views, routes, and a model modification with rich text for feedback. This commit also updates the navigation in the header to include the Awardees page.